requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.4
celery[redis]==5.4.0
redis==5.2.1
sqlalchemy==2.0.36
pydantic-settings==2.7.0
httpx==0.28.1
pydub==0.25.1
python-multipart==0.0.20
aiofiles==24.1.0
sse-starlette==2.2.1
python-dotenv==1.0.1
anthropic==0.42.0
langdetect==1.0.9
deep-translator==1.11.4
faster-whisper==1.2.1
demucs>=4.0.0
pyannote.audio>=3.1
resemblyzer>=0.1.3
openai>=1.0.0
